Fix build problems
