+++ ConsoleApp/Assets/Images/dog_cat.jpg
