+++ ConsoleApp/Assets/Images/kite.jpg
