+++ ConsoleApp/Assets/Images/ski.jpg
