+++ ConsoleApp/Assets/Images/ski2.jpg
