+++ ConsoleApp/ConsoleApp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31515.178
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp", "ConsoleApp.csproj", "{C1F304FC-8214-4358-B2EB-7D7C18A880E5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YOLOv4MLNet", "..\ModelLib\YOLOv4MLNet.csproj", "{AB023224-5845-4BC2-A964-A78F0F9B70DC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C1F304FC-8214-4358-B2EB-7D7C18A880E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1F304FC-8214-4358-B2EB-7D7C18A880E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1F304FC-8214-4358-B2EB-7D7C18A880E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1F304FC-8214-4358-B2EB-7D7C18A880E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AB023224-5845-4BC2-A964-A78F0F9B70DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AB023224-5845-4BC2-A964-A78F0F9B70DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AB023224-5845-4BC2-A964-A78F0F9B70DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AB023224-5845-4BC2-A964-A78F0F9B70DC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A9624771-BB74-49A2-ABD3-B34E5D428681}
+	EndGlobalSection
+EndGlobal
