+++ ModelLib/.gitignore
@@ -134,6 +134,3 @@ $RECYCLE.BIN/
 .DS_Store
 
 _NCrunch*
-/YOLOv4MLNet/Assets/Images
-/.vs
-/ModelLibrary/Assets/Images
