+++ ModelLib/DataStructures/YoloV4BitmapData.cs
