+++ ModelLib/YOLOv4MLNet.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.31515.178
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YOLOv4MLNet", "YOLOv4MLNet.csproj", "{FC5A2373-C004-4E47-8140-F657BAD6D4B8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YoloC", "..\YoloC\YoloC.csproj", "{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,18 +27,6 @@ Global
 		{FC5A2373-C004-4E47-8140-F657BAD6D4B8}.Release|x64.Build.0 = Release|Any CPU
 		{FC5A2373-C004-4E47-8140-F657BAD6D4B8}.Release|x86.ActiveCfg = Release|Any CPU
 		{FC5A2373-C004-4E47-8140-F657BAD6D4B8}.Release|x86.Build.0 = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|x64.Build.0 = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Debug|x86.Build.0 = Debug|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|x64.ActiveCfg = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|x64.Build.0 = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|x86.ActiveCfg = Release|Any CPU
-		{EE5802B9-D7BB-4FCF-B9B3-A208E0C99796}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
